+++ src/components/@index.js
@@ -2,7 +2,6 @@ import AlignContainer from './align_container';
 import Button from './button';
 import Layout from './layout';
 import Poster from './poster';
-import PosterCard from './poster_card';
 import Skeleton from './skeleton';
 
-export { AlignContainer, Button, Layout, Poster, PosterCard, Skeleton };
+export { AlignContainer, Button, Layout, Poster, Skeleton };
