+++ src/pages/@index.js
@@ -1,4 +1,4 @@
-import BoardPage from './board_page';
+import BoardPage from './board_page/board_page';
 import DetailPage from './detail_page';
-import HomePage from './home_page';
+import HomePage from './home_page/home_page';
 export { BoardPage, DetailPage, HomePage };
