+++ src/pages/board_page/board_page.jsx
@@ -2,11 +2,11 @@ import InfiniteScroll from 'react-infinite-scroller';
 import { useInfiniteQuery } from 'react-query';
 import { useParams, useSearchParams } from 'react-router-dom';
 import styled from 'styled-components';
-import AlignContainer from '../components/align_container';
-import movies from '../constants/query_keys/movies';
-import pageNames from '../constants/texts/page_names';
-import moviesFetchFn from '../libs/axios/movie';
-import { color, fontSize, fontWeight } from '../styles/themes/@index';
+import { AlignContainer } from '../../components/@index';
+import movies from '../../constants/query_keys/movies';
+import pageNames from '../../constants/texts/page_names';
+import moviesFetchFn from '../../libs/axios/movie';
+import { color, fontSize, fontWeight } from '../../styles/themes/@index';
 import PostCardList from './components/post_card_list';
 
 /**
